+++ app/src/main/java/com/usal/jorgeav/sportapp/sportselection/SportsListFragment.java
@@ -113,7 +113,7 @@ public class SportsListFragment extends BaseFragment {
 
         String[] sportsNameArray = getResources().getStringArray(R.array.sport_id_values);
         for (String aSportsNameArray : sportsNameArray) {
-            result.add(new Sport(aSportsNameArray, 0f, 1));
+            result.add(new Sport(aSportsNameArray, (double) 0f, 1));
         }
 
         if (getArguments() != null && getArguments().containsKey(BUNDLE_INSTANCE_SPORT_LIST)) {
